created "transformer" package and restructured code
