--- docker_compose_port.go
@@ -1,8 +1,0 @@
-package main
-
-type docker_compose_port struct {
-	Target    string
-	Published string
-	Protocol  string
-	Mode      string
-}
