+++ main.go
@@ -17,13 +17,7 @@ import (
 	"github.com/cloud66/starter/common"
 	"github.com/heroku/docker-registry-client/registry"
 	"github.com/mitchellh/go-homedir"
-
-	"gopkg.in/yaml.v2"
-	"bufio"
-	"unicode"
-	"log"
-	"strconv"
-	"math"
+	"github.com/cloud66/starter/transformer"
 )
 
 type downloadFile struct {
@@ -289,7 +283,7 @@ func main() {
 
 			inFile := string(flagIn)
 			outFile := string(flagOut)
-			ok, _ := transform(inFile, outFile)
+			ok, _ := transformer.Transformer(inFile, outFile, directlyTransformed)
 			if ok == true {
 				common.PrintlnTitle("Done!")
 			}
@@ -339,684 +333,6 @@ func main() {
 	common.PrintlnTitle("Done")
 }
 
-//Transform format functions
-
-type Build_Command struct {
-	Build         Build `yaml:"dockerfile,omitempty"`
-	Build_Command string `yaml:"build,omitempty"`
-}
-
-type Build struct {
-	Dockerfile string `yaml:"dockerfile,omitempty"`
-}
-
-type Deploy struct {
-	Resources Resources `yaml:"resources,omitempty"`
-}
-
-type Resources struct {
-	Limits Limits `yaml:"limits,omitempty"`
-}
-
-type Limits struct {
-	Cpus   string `yaml:"cpus,omitempty"`
-	Memory string `yaml:"memory,omitempty"`
-}
-
-type Command struct {
-	Command []string `yaml:"command,omitempty"`
-}
-
-type Volumes struct {
-	Volumes []string `yaml:"volumes,omitempty"`
-}
-
-type Env_file struct {
-	Env_file []string `yaml:"env_file,omitempty"`
-}
-
-type EnvVars struct {
-	EnvVars []string `yaml:"environment,omitempty"`
-}
-
-type Ports struct {
-	Port        []Port `yaml:"ports,omitempty"`
-	ShortSyntax []string `yaml:"shortsyntax,omitempty"`
-}
-
-type Port struct {
-	Target    string `yaml:"target,omitempty"`
-	Published string `yaml:"published,omitempty"`
-	Protocol  string `yaml:"protocol,omitempty"`
-	Mode      string `yaml:"mode,omitempty"`
-}
-
-func (e *Build_Command) UnmarshalYAML(unmarshal func(interface{}) error) error {
-
-	var build Build
-	err := unmarshal(&build)
-	if err != nil {
-		var single string
-		err := unmarshal(&single)
-		if err != nil {
-			return err
-		}
-		e.Build_Command = single
-	} else {
-		e.Build.Dockerfile = build.Dockerfile
-	}
-	return nil
-}
-
-func (sm *Volumes) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var multi []string
-	err := unmarshal(&multi)
-	if err != nil {
-		var single string
-		err := unmarshal(&single)
-		if err != nil {
-			return err
-		}
-		sm.Volumes = make([]string, 1)
-		sm.Volumes[0] = single
-	} else {
-		sm.Volumes = multi
-	}
-	return nil
-}
-
-func (ef *Env_file) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var multi []string
-	err := unmarshal(&multi)
-	if err != nil {
-		var single string
-		err := unmarshal(&single)
-		if err != nil {
-			return err
-		}
-		ef.Env_file = make([]string, 1)
-		ef.Env_file[0] = single
-	} else {
-		ef.Env_file = multi
-	}
-	return nil
-}
-
-func (sm *EnvVars) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var multi []string
-	err := unmarshal(&multi)
-	if err != nil {
-		var single string
-		err := unmarshal(&single)
-		if err != nil {
-			return err
-		}
-		sm.EnvVars = make([]string, 1)
-		sm.EnvVars[0] = single
-	} else {
-		sm.EnvVars = multi
-	}
-	return nil
-}
-
-//unsupported warnings
-
-func (sm *Command) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var multi []string
-	err := unmarshal(&multi)
-	if err != nil {
-		var single string
-		err := unmarshal(&single)
-		if err != nil {
-			return err
-		}
-		sm.Command = make([]string, 1)
-		sm.Command[0] = single
-	} else {
-		sm.Command = multi
-	}
-	return nil
-}
-
-func (p *Ports) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var multiPorts []Port
-	err := unmarshal(&multiPorts)
-	if err != nil {
-		var singlePort Port
-		err := unmarshal(&singlePort)
-		if err != nil {
-			var multiString []string
-			err := unmarshal(&multiString)
-			if err != nil {
-				var single string
-				err := unmarshal(&single)
-				if err != nil {
-					return err
-				}
-				p.ShortSyntax = make([]string, 1)
-				p.ShortSyntax[0] = single
-			} else {
-				p.ShortSyntax = multiString
-			}
-			return nil
-		}
-		p.Port = make([]Port, 1)
-		p.Port[0] = singlePort
-	} else {
-		p.Port = multiPorts
-	}
-	return nil
-}
-
-type Service struct {
-	Command           Command `yaml:"command,omitempty"`
-	Ports             Ports `yaml:"ports,omitempty"`
-	Build_Command     Build_Command `yaml:"build,omitempty"`
-	Image             string `yaml:"image,omitempty"`
-	Depends_on        []string `yaml:"depends_on,omitempty"`
-	EnvVars           []string `yaml:"environment,omitempty"`
-	Deploy            Deploy `yaml:"deploy,omitempty"`
-	Volumes           Volumes `yaml:"volumes,omitempty"`
-	Stop_grace_period string `yaml:"stop_grace_period,omitempty"`
-	Working_dir       string `yaml:"working_dir,omitempty"`
-	Privileged        bool `yaml:"privileged,omitempty"`
-	Labels            []string `yaml:"labels,omitempty"`
-	Expose            []string `yaml:"expose,omitempty"`
-	Env_file          Env_file `yaml:"env_file,omitempty"`
-
-	//unsupported docker-compose specifications
-	Links          Links
-	Cap_add        Cap_add
-	Cap_drop       Cap_drop
-	Logging        Logging
-	Cgroup_parent  Cgroup_parent
-	Container_name Container_name
-	Devices        Devices
-	Dns            Dns
-	Dns_search     Dns
-	External_links Links
-	Extra_hosts    Extra_hosts
-	Isolation      Isolation
-	Networks       Networks
-	Pid            Extra_hosts
-	Secrets        Secrets
-	Security_opt   Security_opt
-	Userns_mode    Userns_mode
-	Ulimits        Ulimits
-	Healthcheck    Healthcheck
-}
-
-//main transformation format function
-func transform(filename string, formatTarget string) (bool, error) {
-
-	var err error
-	_, err = os.Stat(formatTarget)
-
-	// create file if not exists
-	if os.IsNotExist(err) {
-		var file, err = os.Create(formatTarget)
-		checkError(err)
-		defer file.Close()
-	} else {
-		common.PrintError("File %s already exists. Will be overwritten.\n", formatTarget)
-	}
-
-	//HANDLE ENV_VARS
-	dockerComp, _ := os.Open(filename)
-
-	var dockerLines []string
-
-	scanner := bufio.NewScanner(dockerComp)
-	for scanner.Scan() {
-		dockerLines = append(dockerLines, scanner.Text())
-	}
-
-	var countUpper, countLower, firstNonSpace int
-	var hasCol bool
-	var dockerText string
-
-	initialDocker := dockerLines
-	initialDockerText := ""
-
-	for i := 0; i < len(dockerLines); i++ {
-
-		initialDockerText += initialDocker[i] + "\n"
-
-		countUpper = 0
-		countLower = 0
-		firstNonSpace = -1
-		var j int
-		for j = 0; j < len(dockerLines[i]); j++ {
-			if !unicode.IsSpace(rune(dockerLines[i][j])) {
-				firstNonSpace = j
-				break
-			}
-		}
-		if firstNonSpace > 0 {
-			if dockerLines[i][firstNonSpace] != '-' && !unicode.IsLower(rune(dockerLines[i][firstNonSpace])) {
-
-				for ; j < len(dockerLines[i]); j++ {
-					if unicode.IsUpper(rune(dockerLines[i][j])) {
-						countUpper++
-					}
-					if unicode.IsLower(rune(dockerLines[i][j])) {
-						countLower++
-					}
-					if (dockerLines[i][j] == ':' || dockerLines[i][j] == '=') && countUpper > 1 {
-						if j+1 < len(dockerLines[i]) {
-							if dockerLines[i][j+1] == ' ' {
-								var envKey, envValue string
-								var k int
-								for k = 0; k <= j; k++ {
-									envKey += string(dockerLines[i][k])
-								}
-								for k = k + 1; k < len(dockerLines[i]); k++ {
-									envValue += string(dockerLines[i][k])
-								}
-								dockerLines[i] = envKey + envValue
-							}
-							hasCol = true
-							break
-						}
-					}
-				}
-				if countUpper > 1 && countLower == 0 {
-					hasCol = true
-				}
-				if countUpper > 1 && hasCol && dockerLines[i][firstNonSpace] != '-' {
-					environmentVar := ""
-					for j = firstNonSpace; j < len(dockerLines[i]); j++ {
-						environmentVar += string(dockerLines[i][j])
-					}
-					dockerLines[i] = "     - " + environmentVar
-				}
-			}
-		}
-		dockerText += dockerLines[i] + "\n"
-	}
-	auxFile := filename + "aux"
-
-	dockerComp, _ = os.Create(auxFile)
-	_, err = dockerComp.WriteString(dockerText)
-
-	err = dockerComp.Sync()
-	checkError(err)
-
-	dockerComp.Close()
-
-	yamlFile, err := ioutil.ReadFile(auxFile)
-
-	dockerCompose := Docker_compose{
-		Services: make(map[string]Service),
-		Version:  "",
-	}
-
-	serviceYaml := Serviceyml{
-		Services: make(map[string]common.Service),
-		Dbs:      make([]string, 0),
-	}
-
-	if err := yaml.Unmarshal([]byte(yamlFile), &dockerCompose); err != nil {
-		fmt.Println(err.Error())
-	}
-
-	if len(dockerCompose.Services) == 0 {
-		d := make(map[string]Service)
-		err = yaml.Unmarshal([]byte(yamlFile), &d)
-		checkError(err)
-
-		serviceYaml.Services, serviceYaml.Dbs = copyToServiceYML(d)
-
-	} else {
-
-		serviceYaml.Services, serviceYaml.Dbs = copyToServiceYML(dockerCompose.Services)
-	}
-	if len(serviceYaml.Dbs) != 0 {
-		if serviceYaml.Dbs[len(serviceYaml.Dbs)-1] == "" {
-			serviceYaml.Dbs = serviceYaml.Dbs[:len(serviceYaml.Dbs)-1]
-		}
-	}
-
-	file, err := yaml.Marshal(serviceYaml)
-
-	err = ioutil.WriteFile("service.yml", file, 0644)
-	if err != nil {
-		log.Fatalf("ioutil.WriteFile: %v", err)
-	}
-
-	service_yml, er := os.OpenFile(formatTarget, os.O_RDWR, 0644)
-	checkError(er)
-
-	// write some text to file
-	_, err = service_yml.WriteString(string(file))
-	checkError(err)
-
-	// save changes
-	err = service_yml.Sync()
-	checkError(err)
-
-	service_yml.Close()
-
-	//format long syntax ports
-	service_yml, _ = os.Open(formatTarget)
-	defer service_yml.Close()
-
-	var lines []string
-	scanner = bufio.NewScanner(service_yml)
-	for scanner.Scan() {
-		lines = append(lines, scanner.Text())
-	}
-	text := ""
-	for i := 0; i < len(lines); i++ {
-		if strings.Contains(lines[i], "cpu:") {
-			lines[i] = formatCpu(lines[i])
-		}
-		if strings.Contains(lines[i], "env_vars:") {
-			text += lines[i] + "\n"
-			for i = i + 1; i < len(lines); i++ {
-				if isEnv(lines[i]) {
-					lines[i] = formatEnv_Vars(lines[i])
-					text += lines[i] + "\n"
-				} else {
-					text += lines[i] + "\n"
-					break
-				}
-			}
-		} else {
-			if lines[i] == "    - |-" {
-				lines[i] = "    -"
-			}
-			text += lines[i] + "\n"
-		}
-
-	}
-
-	//write the final service.yml
-	service_yml, _ = os.Create(formatTarget)
-	service_yml, er = os.OpenFile(formatTarget, os.O_RDWR, 0644)
-
-	_, err = service_yml.WriteString(text)
-
-	//rewrite the original docker-compose
-	docker, _ := os.Create(filename)
-	docker, er = os.OpenFile(filename, os.O_RDWR, 0644)
-
-	err = docker.Sync()
-	checkError(err)
-
-	_, er = docker.WriteString(initialDockerText)
-	docker.Close()
-
-	err = os.Remove(auxFile)
-	checkError(err)
-
-	return true, err
-
-}
-
-func copyToServiceYML(d map[string]Service) (map[string]common.Service, []string) {
-
-	serviceYaml := Serviceyml{
-		Services: make(map[string]common.Service),
-		Dbs:      make([]string, 0),
-	}
-	var isDB bool
-	var err error
-	var dbs []string
-
-	directlyTransformed = true
-	for k, v := range d {
-		var current_db string
-		isDB = false
-
-		if v.Image != "" {
-			current_db, isDB = checkDB(v.Image)
-		}
-		if isDB {
-			dbs = append(dbs, current_db)
-		}
-		if !isDB {
-			var longSyntaxPorts []string
-			longSyntaxPorts = v.Expose //expose and long syntax for ports dont work together..i think?
-			if len(v.Ports.ShortSyntax) > 0 {
-				for i := 0; i < len(v.Ports.ShortSyntax); i++ {
-					longSyntaxPorts = append(longSyntaxPorts, v.Ports.ShortSyntax[i])
-				}
-			} else {
-				longSyntaxPorts = []string{}
-				for i := 0; i < len(v.Ports.Port); i++ {
-
-					longSyntax := ""
-					longSyntax = "target: " + v.Ports.Port[i].Target + "\n"
-
-					if v.Ports.Port[i].Protocol == "udp" {
-						longSyntax += "udp: " + v.Ports.Port[i].Published
-						longSyntaxPorts = append(longSyntaxPorts, longSyntax)
-					} else if v.Ports.Port[i].Protocol == "tcp" {
-						reader := bufio.NewReader(os.Stdin)
-						fmt.Printf("\nYou have chosen a TCP protocol for the port published at %s - should it be mapped as HTTP, HTTPS or TCP ? : ", v.Ports.Port[i].Published)
-						var answer string
-						answer, _ = reader.ReadString('\n')
-						answer = strings.ToUpper(answer)
-						if answer == "TCP\n" {
-							longSyntax += "tcp: " + v.Ports.Port[i].Published
-						} else if answer == "HTTP\n" {
-							longSyntax += "http: " + v.Ports.Port[i].Published
-						} else if answer == "HTTPS\n" {
-							longSyntax += "http: " + v.Ports.Port[i].Published
-						}
-						longSyntaxPorts = append(longSyntaxPorts, longSyntax)
-					}
-
-				}
-			}
-
-			var gitPath string
-			gitPath, err = common.GitRootDir("/")
-			//checkError(err)
-			if err != nil {
-
-			}
-			hasGit := common.HasGit(gitPath)
-
-			var gitURL, gitBranch, buildRoot string
-			if hasGit {
-				//common.PrintlnWarning("HAS GIT here!")
-				gitURL = common.RemoteGitUrl(gitPath)
-				gitBranch = common.LocalGitBranch(gitPath)
-				buildRoot, err = common.PathRelativeToGitRoot(gitPath)
-			}
-
-			var serviceYamlService common.Service
-			serviceYamlService.GitRepo = gitURL
-			serviceYamlService.GitBranch = gitBranch
-			serviceYamlService.BuildRoot = buildRoot
-			serviceYamlService.BuildCommand = v.Build_Command.Build_Command
-			serviceYamlService.CommandSlice = v.Command.Command
-			serviceYamlService.Image = v.Image
-			serviceYamlService.Requires = v.Depends_on
-			serviceYamlService.Volumes = v.Volumes.Volumes
-			serviceYamlService.Stop_grace = v.Stop_grace_period
-			serviceYamlService.Work_dir = v.Working_dir
-			serviceYamlService.EnvVarsSlice = v.EnvVars
-			serviceYamlService.Tags = v.Labels
-			serviceYamlService.Dockerfile_path = v.Build_Command.Build.Dockerfile
-			serviceYamlService.Privileged = v.Privileged
-			serviceYamlService.Constraints = common.Constraints{
-				Resources: common.Resources{
-					Memory: v.Deploy.Resources.Limits.Memory,
-					Cpu:    v.Deploy.Resources.Limits.Cpus,
-				},
-			}
-			serviceYamlService.PortsShort = longSyntaxPorts
-
-			if v.Env_file.Env_file != nil {
-				var lines []string
-				for i := 0; i < len(v.Env_file.Env_file); i++ {
-					lines = readEnv_file(v.Env_file.Env_file[i])
-					for j := 0; j < len(lines); j++ {
-						serviceYamlService.EnvVarsSlice = append(serviceYamlService.EnvVarsSlice, lines[j])
-					}
-				}
-			}
-
-			//serviceYamlService.EnvVarsSlice = formatEnv_Vars(serviceYamlService.EnvVarsSlice)
-
-			if serviceYamlService.Image != "" {
-				serviceYamlService.GitRepo = ""
-				serviceYamlService.GitBranch = ""
-				serviceYamlService.BuildRoot = ""
-			}
-			// assign stuff via v, k is the service name
-			serviceYaml.Services[k] = serviceYamlService
-
-		}
-
-	}
-	return serviceYaml.Services, dbs
-}
-
-func formatCpu(cpuString string) string {
-	var i, auxInt, p int
-	var auxString string
-	//common.PrintlnTitle(cpuString)
-	for i = 0; i < len(cpuString); i++ {
-		if cpuString[i] == '\'' || cpuString[i] == '"' {
-			p = i
-			break
-		}
-	}
-	for i = i + 1; i < len(cpuString); i++ {
-		if cpuString[i] == '\'' || cpuString[i] == '"' {
-			break
-		} else {
-			auxString += string(cpuString[i])
-		}
-	}
-	auxFloat, err := strconv.ParseFloat(auxString, 64)
-	checkError(err)
-	fract := auxFloat - math.Floor(auxFloat)
-	if auxFloat < 1 {
-		auxInt = 1
-	} else if fract < 0.5 {
-		auxInt = int(math.Floor(auxFloat))
-	} else {
-		auxInt = int(math.Ceil(auxFloat))
-	}
-	cpuString = cpuString[:p] + strconv.Itoa(auxInt)
-	return cpuString
-}
-
-func isEnv(line string) bool {
-	for i := 0; i < len(line); i++ {
-		if !unicode.IsSpace(rune(line[i])) {
-			if line[i] == '-' {
-				return true
-			}
-			break
-		}
-	}
-	return false
-}
-
-func formatEnv_Vars(env string) string {
-	var j int
-	for j = 0; j < len(env); j++ {
-		if env[j] == '-' {
-			env = env[:j] + " " + env[j+1:]
-		}
-	}
-	for ; j < len(env); j++ {
-		if env[j] == '\'' {
-			env = env[:j] + " " + env[j+1:]
-		}
-		if !unicode.IsSpace(rune(env[j])) {
-			break
-		}
-	}
-	for j = 0; j < len(env); j++ {
-		if j+1 < len(env) {
-			if env[j] == ':' {
-				env = env[:j+1] + " " + env[j+1:]
-				break
-			}
-			if env[j] == '=' {
-				env = env[:j] + ": " + env[j+1:]
-				break
-			}
-		}
-
-	}
-	for j = len(env) - 1; j >= 0; j-- {
-		if env[j] == '-' {
-			env = env[:j] + " " + env[j+1:]
-		}
-	}
-
-	return env
-}
-
-func readEnv_file(path string) []string {
-	var lines []string
-	var env_vars []string
-
-	envFile, err := os.Open(path)
-	checkError(err)
-
-	scanner := bufio.NewScanner(envFile)
-	for scanner.Scan() {
-		lines = append(lines, scanner.Text())
-	}
-
-	for i := 0; i < len(lines); i++ {
-
-		if !isCommentLine(lines[i]) {
-			env_vars = append(env_vars, lines[i])
-		}
-	}
-	envFile.Close()
-	return env_vars
-}
-
-func isCommentLine(line string) bool {
-	var i int
-	for i = 0; i < len(line); i++ {
-		if !unicode.IsSpace(rune(line[i])) {
-			break
-		}
-	}
-	if line != "" {
-		if line[i] == '#' {
-			return true
-		}
-	}
-	return false
-}
-
-func checkDB(image string) (string, bool) {
-	db_check := []string{"mysql", "postgresql", "redis", "mongodb", "elasticsearch", "glusterfs", "influxdb", "rabbitmq", "sqlite"}
-	var prefix string
-	if len(image) < 5 {
-		prefix = image
-	} else {
-		for i := 0; i < 4; i++ {
-			prefix += string(image[i])
-		}
-	}
-	for i := 0; i < len(db_check); i++ {
-		if strings.Contains(image, db_check[i]) || strings.Contains(db_check[i], image) || strings.Contains(prefix, db_check[i]) || strings.Contains(db_check[i], prefix) {
-			return db_check[i], true
-		}
-	}
-	return "", false
-}
-
-func checkError(err error) {
-	if err != nil {
-		fmt.Println(err.Error())
-		os.Exit(0)
-	}
-}
 
 func analyze(
 	updateTemplates bool,
@@ -1068,6 +384,7 @@ func analyze(
 		dockerComposeYAMLTemplateDir = templates
 	}
 
+	//look for docker-compose file in order to transform directly to service.yml
 	dockercomposePath := filepath.Join(path, "docker-compose.yml")
 	serviceYAMLPath := filepath.Join(path, "service.yml")
 	if _, err := os.Stat(serviceYAMLPath); err == nil {
@@ -1077,7 +394,7 @@ func analyze(
 		} else {
 			if _, err := os.Stat(dockercomposePath); err == nil {
 				// docker-compose.yml exists
-				_, err = transform(dockercomposePath, "service.yml")
+				directlyTransformed, err = transformer.Transformer(dockercomposePath, "service.yml", directlyTransformed)
 				return nil, err
 			}
 		}
@@ -1086,7 +403,7 @@ func analyze(
 
 		if _, err := os.Stat(dockercomposePath); err == nil {
 			// docker-compose.yml exists
-			_, err = transform(dockercomposePath, "service.yml")
+			directlyTransformed, err = transformer.Transformer(dockercomposePath, "service.yml", directlyTransformed)
 			return nil, err
 		}
 	}
